feat: add alert for group spending per block
Enabled when `alert_webhook` and `alert_group_spend_sats` are set in the
config file. If any group spends more than `alert_group_spend_sats` in
one block. The last block that was checked is updated in the field
`last_alert_block` so that the alert is only sent once.